+++ ergo_std/ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 * Add `std:io:is-terminal` to check whether a stream is connected to a terminal.
+* Add `std:fs:file-type` to get the file type of a path.
 
 ## 1.0.0-rc.1  -- 2021-12-08
 ### New Features